requirements.txt
# ==============================
# LIAM AI Assistant Requirements
# ==============================
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyAudio>=0.2.13
elevenlabs>=0.2.24  # For premium voice capabilities
sounddevice>=0.4.6  # Audio playback
soundfile>=0.12.1   # Sound file manipulation
pywin32>=306        # Windows API access (includes win32gui, win32con, win32api, win32clipboard)
pyautogui>=0.9.54   # Fallback GUI automation
pyperclip>=1.8.2    # Enhanced clipboard operations
opencv-python>=4.8.1.78
Pillow>=10.1.0
numpy>=1.26.2
tensorflow>=2.15.0  # ML framework
tf-keras>=2.15.0    # Keras integration for TF
keras>=2.15.0       # High-level neural networks API
deepface>=0.0.79    # Face recognition
retina-face>=0.0.13 # Precise face detection
mtcnn>=0.1.1        # Multi-task CNN for face detection
python-xlib>=0.33; platform_system=="Linux"  # Linux GUI automation
setuptools>=65.5.0
wheel>=0.38.4
tqdm>=4.66.1        # Progress bars
psutil>=5.9.0       # System monitoring
pipwin>=0.5.2       # Helps with PyAudio installation on Windows
pytesseract>=0.3.10   # For OCR

